requirements.txt
wandb==0.21.3
fastapi[standard]==0.116.1
torch==2.8.0
torchvision==0.23.0
loadotenv==1.0.1


